+++ backend/src/main/java/ee/hor/tablebooking/service/TableService.java
@@ -2,9 +2,12 @@ package ee.hor.tablebooking.service;
 
 import ee.hor.tablebooking.dto.TableDto;
 import ee.hor.tablebooking.entity.TableEntity;
+import ee.hor.tablebooking.excpetion.EntityInUseException;
+import ee.hor.tablebooking.excpetion.ResourceNotFoundException;
 import ee.hor.tablebooking.mapper.TableMapper;
 import ee.hor.tablebooking.repository.TableRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,8 +20,9 @@ public class TableService {
     private final TableMapper tableMapper;
 
     public TableDto getTable(UUID id) {
-        // FIXME: throw error if not found
-        TableEntity tableEntity = tableRepository.findById(id).orElse(null);
+        TableEntity tableEntity = tableRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Table with given id does not exist")
+        );
 
         return tableMapper.mapToDto(tableEntity);
     }
@@ -38,6 +42,14 @@ public class TableService {
     }
 
     public void deleteTable(UUID id) {
-        tableRepository.deleteById(id);
+        if (!tableRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Table with given id does not exist");
+        }
+
+        try {
+            tableRepository.deleteById(id);
+        } catch (DataIntegrityViolationException _) {
+            throw new EntityInUseException("Table is being used by other objects, remove them before proceeding with deletion");
+        }
     }
 }
